Fix minor issues with fonts
This fixes #7, #8, and #9 by modifying the glyphs to remove the
extraneous pixels within them. Note that the modification to Page.pm
was required to make the pages identical to the ones with the previous
fonts. The reason is that when the fonts are re-exported, the values for
char_up and char_down change in a way that is inconsistent with the rest
of the files.

As a note to future self, I tried fixing this within the font itself by
dumping the hhea table and os2 tables and comparing them and/or merging
them, but this unfortunately didn't work. Also note that in most cases,
this is not a real problem (ex the web handles these well without a problem).